Separate the implementation of blob/queue/table into separate subfolders.
