+++ azure-storage/azure/storage/queue/__init__.py
@@ -12,38 +12,17 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #--------------------------------------------------------------------------
-from .constants import (
-    __author__,
-    __version__,
-    X_MS_VERSION,
-    DEV_ACCOUNT_NAME,
-    DEV_ACCOUNT_KEY,
-    DEFAULT_HTTP_TIMEOUT,
+from ..constants import (
+    QUEUE_SERVICE_HOST_BASE,
+    DEV_QUEUE_HOST,
 )
 
 from .models import (
-    RetentionPolicy,
-    Logging,
-    HourMetrics,
-    MinuteMetrics,
-    StorageServiceProperties,
-    AccessPolicy,
-    SignedIdentifier,
-    SignedIdentifiers,
+    QueueEnumResults,
+    Queue,
+    QueueMessagesList,
+    QueueMessage,
+    QueueSharedAccessPermissions,
 )
 
-from .cloudstorageaccount import CloudStorageAccount
-from .sharedaccesssignature import (
-    SharedAccessSignature,
-    SharedAccessPolicy,
-)
-from .auth import (
-    _StorageSharedKeyAuthentication,
-    StorageNoAuthentication,
-    StorageSASAuthentication,
-    StorageSharedKeyAuthentication,
-    StorageTableSharedKeyAuthentication,
-)
-from .connection import (
-    StorageConnectionParameters,
-)
+from .queueservice import QueueService
