+++ tests/test_storage_queue.py
@@ -23,6 +23,8 @@ from azure.storage import (
     SharedAccessPolicy,
     SignedIdentifier,
     SignedIdentifiers,
+)
+from azure.storage.queue import (
     QueueService,
     QueueSharedAccessPermissions,
 )
